Getting Started

Installation

Clone the repository and add the package to your Julia environment:

using Pkg
Pkg.add("RxInferServer")

Refer to Julia documentation for more information on how to add packages to your Julia environment: Julia Documentation.

Starting the Server

RxInferServer.serve — Function

serve() -> HTTP.Server

Start the RxInfer API server with the configured settings. Official documentation is available at https://server.rxinfer.com/

Description

Initializes and starts an HTTP server that exposes RxInfer functionality through a REST API. The server uses the OpenAPI specification defined in RxInferServerOpenAPI module to register endpoints.

Features

Configurable port via the RXINFER_SERVER_PORT environment variable (default: 8000)

Graceful shutdown with proper resource cleanup when interrupted

Loads the .env files based on the RXINFER_SERVER_ENV environment variable

When Revise.jl is loaded in the current Julia session, and the RXINFER_SERVER_ENABLE_HOT_RELOAD environment variable is set to "true", the server will hot reload the source code and models when the source code changes.

This is a blocking operation that runs until interrupted (e.g., with Ctrl+C). To gracefully shut down the server, type the q or quit and press ENTER in the REPL. Note that Ctrl+C cannot be catched reliably when running in a non-interactive session.

Examples

using RxInferServer

# Start the server with default settings (blocks until interrupted)
RxInferServer.serve()

See Also

RxInferServer.RXINFER_SERVER_ENV: The environment on which the RxInfer server will run, determines which .env files are loaded

RxInferServer.RXINFER_SERVER_PORT: The port on which the RxInfer server will run

RxInferServer.RXINFER_SERVER_ENABLE_HOT_RELOAD: Check if hot reloading is enabled

RxInferServer.RXINFER_SERVER_SHOW_BANNER: Whether to show the welcome banner

RxInferServer.RXINFER_SERVER_LISTEN_KEYBOARD: Whether to listen for keyboard input to quit the server

To start the server with default settings:

using RxInferServer

# This will block until stopped manually
RxInferServer.serve()

Note

When running the server from script, e.g. make serve, Ctrl-C might not work properly. See How do I catch CTRL-C in a script? for more information.

To start the server in the background:

@async RxInferServer.serve()

# or 

Threads.@spawn RxInferServer.serve()

Read more about the @async and Threads.@spawn macros in the Julia Documentation.

Where to go next?

Read more about the different configuration options in the Configuration section.

If you want to contribute to the project, read more about the Development section.